Times -> NumberOfTimes
